More docs; check_status changes; split up src; TypeErrors
 - suggested "".format() applied but doesn't affect pylint output for me locally.

 - Split code into 3 additional files largely based on classes. This is primarily to reduce required memory footprint as much as possible, but also for improved code clarity. Now have:

    __init__: EMC2101 main class. Could require "CV" classes but only if used.
    emc2101_enums: The "CV" classes formerly in __init__
    emc2101_fanspeed: The FanSpeedLUT class formerly in emc2101_lut
    emc2101_ext: The extended version of EMC2101, implementing status checks.
    emc2101_lut: The LUT version of EMC2101_EXT, adding support for LUT.
    emc2101_regs: Constants, mostly register addresses but some others now.

 - Removed check_status code from EMC2101, so that class never does checks, though devstatus property makes it possible manually.

 - Converted EMC2101_EXT to include such checks but only when enabled first (`auto_check_status` attr), and included `last_status` property to read last hardware value (because chip zeroes reg out on read).

 - Lots and lots of docstrings updated and improved.

 - A couple of exception renamings (ValueError to TypeError).

 - Added Makefile to compile the python source to mpy code and arm native code, for size checks. Can remove later if wanted.
